@@ -44,9 +44,10 @@ using namespace std;
 #include "FeatureQueue.h"
 #include "SpeechWrap.h"
 #include <Audio.h>
+#include "resample.h"
 #include "Model.h"
 #include "paraformer_onnx.h"
-#include "librapidasrapi.h"
+#include "libfunasrapi.h"
 
 
 using namespace paraformer;
