@@ -109,7 +109,7 @@ class ASRServicer(paraformer_pb2_grpc.ASRServicer):
                             else:
                                 asr_result = ""
                         elif self.backend == "onnxruntime":
-                            from rapid_paraformer.utils.frontend import load_bytes
+                            from funasr_onnx.utils.frontend import load_bytes
                             array = load_bytes(tmp_data)
                             asr_result = self.inference_16k_pipeline(array)[0]
                         end_time = int(round(time.time() * 1000))
